+++ source/B2B.Transactions.Api/Middleware/ServiceBus/ServiceBusMessageMetadata.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace B2B.Transactions.Infrastructure.Configuration.ServiceBus
+namespace B2B.Transactions.Api.Middleware.ServiceBus
 {
     public class ServiceBusMessageMetadata
     {
