+++ source/B2B.Transactions.Api/Program.cs
@@ -18,8 +18,8 @@ using System.Threading.Tasks;
 using B2B.Transactions.Api.Middleware.Authentication.Bearer;
 using B2B.Transactions.Api.Middleware.Authentication.MarketActors;
 using B2B.Transactions.Api.Middleware.Correlation;
+using B2B.Transactions.Api.Middleware.ServiceBus;
 using B2B.Transactions.Infrastructure.Configuration;
-using B2B.Transactions.Infrastructure.Configuration.ServiceBus;
 using Energinet.DataHub.Core.Logging.RequestResponseMiddleware;
 using Microsoft.Extensions.Hosting;
 using Microsoft.IdentityModel.Protocols;
