+++ source/B2B.Transactions.CimMessageAdapter/MessageReceiver.cs
@@ -24,7 +24,7 @@ using B2B.Transactions;
 using B2B.Transactions.Authentication;
 using B2B.Transactions.Messages;
 using B2B.Transactions.Transactions;
-using B2B.Transactions.Xml.Incoming;
+using B2B.Transactions.Xml.SchemaStore;
 
 namespace B2B.CimMessageAdapter
 {
