+++ source/B2B.Transactions.CimMessageAdapter/Messages/MessageParser.cs
@@ -20,7 +20,7 @@ using System.Xml;
 using System.Xml.Schema;
 using B2B.CimMessageAdapter.Errors;
 using B2B.Transactions.Messages;
-using B2B.Transactions.Xml.Incoming;
+using B2B.Transactions.Xml.SchemaStore;
 
 namespace B2B.CimMessageAdapter.Messages
 {
