+++ source/B2B.Transactions.Infrastructure/Configuration/CompositionRoot.cs
@@ -29,8 +29,8 @@ using B2B.Transactions.Infrastructure.Serialization;
 using B2B.Transactions.Infrastructure.Transactions;
 using B2B.Transactions.OutgoingMessages;
 using B2B.Transactions.Transactions;
-using B2B.Transactions.Xml.Incoming;
 using B2B.Transactions.Xml.Outgoing;
+using B2B.Transactions.Xml.SchemaStore;
 using Energinet.DataHub.Core.Logging.RequestResponseMiddleware.Storage;
 using Energinet.DataHub.MarketRoles.Domain.SeedWork;
 using EntityFrameworkCore.SqlServer.NodaTime.Extensions;
