+++ source/B2B.Transactions.IntegrationTests/Infrastructure/OutgoingMessages/MessagePublisherTests.cs
@@ -20,8 +20,8 @@ using B2B.Transactions.IntegrationTests.Fixtures;
 using B2B.Transactions.IntegrationTests.TestDoubles;
 using B2B.Transactions.IntegrationTests.Transactions;
 using B2B.Transactions.OutgoingMessages;
-using B2B.Transactions.Xml.Incoming;
 using B2B.Transactions.Xml.Outgoing;
+using B2B.Transactions.Xml.SchemaStore;
 using Energinet.DataHub.MarketRoles.Domain.SeedWork;
 using Energinet.DataHub.MessageHub.Model.Model;
 using Xunit;
