+++ source/B2B.Transactions/Xml/Outgoing/AcceptMessageFactory.cs
@@ -20,7 +20,6 @@ using System.Xml.Linq;
 using System.Xml.Schema;
 using B2B.Transactions.OutgoingMessages;
 using B2B.Transactions.Transactions;
-using B2B.Transactions.Xml.Incoming;
 using Energinet.DataHub.MarketRoles.Domain.SeedWork;
 
 namespace B2B.Transactions.Xml.Outgoing
