+++ source/B2B.Transactions/Xml/SchemaStore/ISchemaProvider.cs
@@ -15,7 +15,7 @@
 using System.Threading.Tasks;
 using System.Xml.Schema;
 
-namespace B2B.Transactions.Xml.Incoming
+namespace B2B.Transactions.Xml.SchemaStore
 {
     /// <summary>
     /// Provides XML schemas for CIM messages
