+++ source/B2B.Transactions/Xml/SchemaStore/SchemaStore.cs
@@ -15,7 +15,7 @@
 using System.Collections.Generic;
 using System.IO;
 
-namespace B2B.Transactions.Xml.Incoming
+namespace B2B.Transactions.Xml.SchemaStore
 {
     public class SchemaStore
     {
